Move common package up 1 level
