+++ src/main/java/dev/nincodedo/ninbot/NinbotRunner.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot;
 
-import dev.nincodedo.ninbot.components.common.Schedulable;
+import dev.nincodedo.ninbot.common.Schedulable;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.entities.Activity;
 import net.dv8tion.jda.api.entities.Guild;
