+++ src/main/java/dev/nincodedo/ninbot/common/message/Impersonation.java
@@ -1,4 +1,4 @@
-package dev.nincodedo.ninbot.components.common.message;
+package dev.nincodedo.ninbot.common.message;
 
 public record Impersonation(String name, String iconUrl) {
     public static Impersonation of(String name, String iconUrl) {
