+++ src/main/java/dev/nincodedo/ninbot/components/channel/TempVoiceChannelManager.java
@@ -1,7 +1,7 @@
 package dev.nincodedo.ninbot.components.channel;
 
-import dev.nincodedo.ninbot.components.common.Emojis;
-import dev.nincodedo.ninbot.components.common.StatAwareListenerAdapter;
+import dev.nincodedo.ninbot.common.Emojis;
+import dev.nincodedo.ninbot.common.StatAwareListenerAdapter;
 import dev.nincodedo.ninbot.components.config.component.ComponentService;
 import dev.nincodedo.ninbot.components.config.component.ComponentType;
 import dev.nincodedo.ninbot.components.stats.StatManager;
