+++ src/main/java/dev/nincodedo/ninbot/components/command/AbstractCommand.java
@@ -1,10 +1,10 @@
 package dev.nincodedo.ninbot.components.command;
 
-import dev.nincodedo.ninbot.components.common.Constants;
-import dev.nincodedo.ninbot.components.common.LocaleService;
-import dev.nincodedo.ninbot.components.common.RolePermission;
-import dev.nincodedo.ninbot.components.common.message.MessageAction;
-import dev.nincodedo.ninbot.components.common.message.WebhookHelper;
+import dev.nincodedo.ninbot.common.Constants;
+import dev.nincodedo.ninbot.common.LocaleService;
+import dev.nincodedo.ninbot.common.RolePermission;
+import dev.nincodedo.ninbot.common.message.MessageAction;
+import dev.nincodedo.ninbot.common.message.WebhookHelper;
 import dev.nincodedo.ninbot.components.config.ConfigService;
 import dev.nincodedo.ninbot.components.stats.StatManager;
 import lombok.Data;
