+++ src/main/java/dev/nincodedo/ninbot/components/event/EventCommandName.java
@@ -1,0 +1,7 @@
+package dev.nincodedo.ninbot.components.event;
+
+import dev.nincodedo.ninbot.common.CommandNameEnum;
+
+public enum EventCommandName implements CommandNameEnum {
+    LIST, PLAN
+}
