+++ src/main/java/dev/nincodedo/ninbot/components/stream/StreamListener.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.components.stream;
 
-import dev.nincodedo.ninbot.components.common.StatAwareListenerAdapter;
+import dev.nincodedo.ninbot.common.StatAwareListenerAdapter;
 import dev.nincodedo.ninbot.components.config.ConfigConstants;
 import dev.nincodedo.ninbot.components.config.ConfigService;
 import dev.nincodedo.ninbot.components.config.component.ComponentService;
