+++ src/main/java/dev/nincodedo/ninbot/components/tally/TallyCommandName.java
@@ -1,0 +1,7 @@
+package dev.nincodedo.ninbot.components.tally;
+
+import dev.nincodedo.ninbot.common.CommandNameEnum;
+
+public enum TallyCommandName implements CommandNameEnum {
+    ADD, GET, SUBTRACT
+}
